ps y cambios en el .gitignore

+++ Kernel/include/keyboard.h
@@ -7,7 +7,7 @@
 void initKeyboard();
 
 //recibe el scancode del teclado y escribe el codigo ascii en el pipe del teclado
-void keyboardHandler(uint64_t keyPressed);
+void keyboardHandler(uint8_t keyPressed);
 
 //lee un caracter ascii desde el pipe del teclado
 char readKeyboard();
